arm: dts: sun8i: a83t: Set timer node to use phy timer
By default, virtual timers are used. These timers need an offset
that must be set by firmware, for example. In case of SMP support,
after a reset, this offset is in "unknown" state and produced
a hang of the kernel.

Use "arm,cpu-registers-not-fw-configured" property allows to use
physical timers instead of virtual ones.

Signed-off-by: Mylène Josserand <mylene.josserand@free-electrons.com>